+++ bot_config.example.py
@@ -1,4 +1,4 @@
-# ВЕРСИЯ СКРИПТА v1.686
+# ВЕРСИЯ СКРИПТА v1.687
 
 token = 'MyBotFatherToken'  # ключ api бота
 usernames = ['MyTelegramLogin']  # Ваш логин в телеграмме без @, не бота.